backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.11
redis==6.4.0
celery==5.5.3
django-cors-headers==4.9.0
django-filter==25.2
Pillow==12.0.0